refactor(validators): extract dispatch-entry predicates; retire test_dispatch_registry (#1207 B5)
Lift check_entry_shape/check_composite_key_exceptional into
planner/validators/_dispatch_registry.py and repoint the E006 smoke importer.
Enforcement already lives in conventions/schema/test_dispatch_map_is_registry.

Refs #1385
